Separate out LoopScope, to be passed around as an object
